Address feedback: extract SystemInfoCommand.cpp, remove Flush
